relaxed/flexible piepline (model)
modules
- pred: LRZ, LRZ-ZZ, LRZ-P, SPL
- hist: SP, generic, (null)
- codec: HF, HFR, FZGPU's, RLE

The refactorizaiton is done to select modules in the pipeline flexibly, with the constrains/limitation as follows. 1) Hist-SP relies on the distribution pattern; with ZigZag'ed input, the unmodified Hist-SP may not result in the optimal speed, such that Hist-generic is the safe fallback. 2) Hist-SP are to be modified to fit the ZigZag pattern. 3) Not all codec-1's need a histogram, such that Hist-(null) is reserved. 4) LRZ and LRZ-ZZ are almost identical, except for the ZigZag'ed output. 5) LRZ-P has not yet enabled LRZ-ZZ backport.